@@ -13,26 +13,27 @@
   },
   "devDependencies": {
     "@sveltejs/adapter-auto": "^3.0.0",
+    "@sveltejs/adapter-vercel": "^5.3.2",
     "@sveltejs/kit": "^2.0.0",
     "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@tailwindcss/typography": "^0.5.13",
     "@types/eslint": "^8.56.7",
+    "autoprefixer": "^10.4.19",
     "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-svelte": "^2.36.0",
     "globals": "^15.0.0",
+    "postcss": "^8.4.38",
     "prettier": "^3.1.1",
     "prettier-plugin-svelte": "^3.1.2",
-    "svelte": "^4.2.7",
+    "prettier-plugin-tailwindcss": "^0.6.4",
+    "svelte": "^5.0.0-next.160",
     "svelte-check": "^3.6.0",
+    "tailwindcss": "^3.4.4",
     "tslib": "^2.4.1",
     "typescript": "^5.0.0",
     "typescript-eslint": "^8.0.0-alpha.20",
-    "vite": "^5.0.3",
-    "tailwindcss": "^3.4.4",
-    "postcss": "^8.4.38",
-    "autoprefixer": "^10.4.19",
-    "@tailwindcss/typography": "^0.5.13",
-    "prettier-plugin-tailwindcss": "^0.6.4"
+    "vite": "^5.0.3"
   },
   "type": "module",
   "dependencies": {
